% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRecode.R
\name{addRecode}
\alias{addRecode}
\title{Update Specification with Recoded Variables}
\usage{
addRecode(dd, dat, vars)
}
\arguments{
\item{dd}{Specification list; this describes the desired state of the data set.}

\item{dat}{data.frame to be modified; when missing will be determined from specification.}

\item{vars}{character vector or list; column names that should be recoded. If in list form, elements
should be individual values to recode for given column names.}
}
\value{
Updated specification list
}
\description{
Create a specification list with \code{datadict}. Then apply this
function to prepare for recoded variables.
}
\examples{
data(airquality)
spec <- datadict(airquality)
upspec <- addRecode(spec, vars = list(Month = c(5,6,7,8,9)))

}
