% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadict.R
\name{datadict}
\alias{datadict}
\title{Create Specification List}
\usage{
datadict(dat)
}
\arguments{
\item{dat}{data.frame}
}
\value{
list
}
\description{
A specification list contains information about a data set such as variables and its labels.
It can be modified and passed to \code{transform} to make changes to the data.
}
\examples{
data(ChickWeight)
spec <- datadict(ChickWeight)

}
