% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastate-package.R
\docType{package}
\name{datastate-package}
\alias{datastate}
\alias{datastate-package}
\title{data.frame State Description and Transformation}
\description{
Describe a data.frame object and provide utilities to
modify state from a YAML file
}
\author{
\strong{Maintainer}: Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})

}
