% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{dd2yaml}
\alias{dd2yaml}
\alias{yaml2dd}
\title{Turn Specification into YAML}
\usage{
dd2yaml(dd, file = stdout())

yaml2dd(file)
}
\arguments{
\item{dd}{Specification list; this describes the desired state of the data set.}

\item{file}{YAML file name; defaults to standard output.}
}
\description{
Convert a given a specification list into YAML.
}
\examples{
data(ChickWeight)
spec <- datadict(ChickWeight)
spec$variables[[1]]$units <- 'gm'
spec$variables[[2]]$label <- 'time since birth'
spec$variables[[2]]$units <- 'days'
spec$variables[[4]]$label <- 'protein diet'
spec$variables[[3]] <- NULL
dd2yaml(spec)

}
