% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFactor.R
\name{makeFactor}
\alias{makeFactor}
\alias{makeOrderedFactor}
\title{Update Specification with Factor Variables}
\usage{
makeFactor(dd, dat, vars, ordered = FALSE)
}
\arguments{
\item{dd}{Specification list; this describes the desired state of the data set.}

\item{dat}{data.frame to be modified; when missing will be determined from specification.}

\item{vars}{character vector; column names that should be considered categorical.}

\item{ordered}{logical flag to determine if levels should be regarded as ordered. See \code{\link[base]{factor}}.}
}
\value{
Updated specification list
}
\description{
Create a specification list with \code{datadict}. Then apply this
function for proposed changes to factor variables.
}
\examples{
data(airquality)
spec <- datadict(airquality)
upspec <- makeFactor(spec, vars = 'Month')

}
