% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mod}
\alias{mod}
\title{Transform data.frame with Given Specifications}
\usage{
mod(dd, dat, delete = TRUE, modify = TRUE, codefile = "", Hmisc_label = TRUE)
}
\arguments{
\item{dd}{Specification list; this describes the desired state of the data set.}

\item{dat}{data.frame to be modified; when missing will be determined from specification.}

\item{delete}{When TRUE data.frame variables not found in the specification will be removed.}

\item{modify}{Set to FALSE for trial run, and the data set will not be modified.}

\item{codefile}{file name to log code and comments with requested modifications; defaults to standard output.}

\item{Hmisc_label}{Label variables with \code{\link[Hmisc]{label}} if available.}
}
\value{
The modified data.frame, or original returned invisibly.
}
\description{
Create a specification list with \code{datadict} in order to make modifications to a data set.
}
\examples{
data(ChickWeight)
spec <- datadict(ChickWeight)
spec$variables[[1]]$label <- 'body weight'
spec$variables[[1]]$units <- 'gm'
spec$variables[[2]]$label <- 'time since birth'
spec$variables[[2]]$units <- 'days'
spec$variables[[3]]$label <- 'chick identifier'
spec$variables[[4]]$label <- 'protein diet'
modCW <- mod(spec, Hmisc_label = FALSE)

}
