% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varNames.R
\name{varNames}
\alias{varNames}
\title{Variables Names in Given Specifications}
\usage{
varNames(dd)
}
\arguments{
\item{dd}{Specification list; this describes the desired state of the data set.}
}
\value{
character vector
}
\description{
Given a specification list, return the variable names.
}
\examples{
data(ChickWeight)
spec <- datadict(ChickWeight)
varNames(spec)
spec$variables[[3]] <- NULL
varNames(spec)

}
